2019.09.15

+++ src/main/java/com/like/hrm/appointment/service/AppointmentService.java
@@ -1,11 +1,15 @@
 package com.like.hrm.appointment.service;
 
+import javax.persistence.EntityNotFoundException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.like.hrm.appointment.boundary.AppointmentCodeDTO;
+import com.like.hrm.appointment.boundary.DeptTypeDTO;
+import com.like.hrm.appointment.boundary.JobTypeDTO;
 import com.like.hrm.appointment.domain.event.ProcessEvent;
 import com.like.hrm.appointment.domain.model.AppointmentCode;
 import com.like.hrm.appointment.domain.model.AppointmentCodeDetail;
@@ -37,7 +41,7 @@ public class AppointmentService {
 		return appointmentRepository.getAppointmentCode(codeId);
 	}
 	
-	public void saveAppointmentCode(AppointmentCodeDTO.CodeSave dto) {
+	public void saveAppointmentCode(AppointmentCodeDTO.SaveCode dto) {
 		AppointmentCode appointmentCode = appointmentRepository.getAppointmentCode(dto.getCode());
 		
 		if (appointmentCode == null ) {		
@@ -57,11 +61,13 @@ public class AppointmentService {
 		
 	public void saveAppointmentCodeDetail(AppointmentCodeDTO.SaveCodeDetail dto) {
 		AppointmentCode appointmentCode = appointmentRepository.getAppointmentCode(dto.getCode());			
-		AppointmentCodeDetail appointmentCodeDetail = appointmentCode.getCodeDetail(dto.getPkCodeDetails());
 		
 		if (appointmentCode == null) {
-			throw new IllegalArgumentException(dto.getCode() + " 엔티티가 존재하지 않습니다.");
-		}		
+			throw new EntityNotFoundException(dto.getCode() + " 엔티티가 존재하지 않습니다.");
+		}
+		
+		AppointmentCodeDetail appointmentCodeDetail = appointmentCode.getCodeDetail(dto.getPkCodeDetails());		
+				
 		
 		if (appointmentCodeDetail == null) {
 			appointmentCodeDetail = dto.newAppointmentCodeDetail(appointmentCode);
@@ -77,27 +83,61 @@ public class AppointmentService {
 		appointmentCode.deleteAppointmentCodeDetail(dto.getPkCodeDetails());			
 	}
 		
-	public DeptType getDeptType(String id) {
-		return appointmentRepository.getDeptType(id);
+	public DeptType getDeptType(String code) {
+		return appointmentRepository.getDeptType(ChangeType.DEPT.toString() + code);
 	}
 	
-	public void saveDeptType(DeptType deptType) {
+	public void saveDeptType(DeptTypeDTO.SaveCode dto) {
+		DeptType deptType = appointmentRepository.getDeptType(ChangeType.DEPT.toString() + dto.getCode());
+		
+		if (deptType == null) {
+			deptType = dto.newDeptType();
+		} else {					
+			deptType.changeInfo(dto.getCodeName()
+					 		   ,dto.isUseYn()
+					 		   ,dto.getSequence()
+					 		   ,dto.getComment());
+		}
+		
 		appointmentRepository.saveDeptType(deptType);		
 	}
 
-	public void deleteDeptType(DeptType deptType) {
-		appointmentRepository.deleteDeptType(deptType);		
+	public void deleteDeptType(String code) {
+		DeptType deptType = appointmentRepository.getDeptType(ChangeType.DEPT.toString() + code);
+		
+		if (deptType == null) {
+			throw new EntityNotFoundException(code + " 엔티티가 존재하지 않습니다.");
+		}
+		
+		appointmentRepository.deleteDeptType(deptType);					
 	}
 
-	public JobType getJobType(String id) {
-		return appointmentRepository.getJobType(id);
+	public JobType getJobType(String code) {
+		return appointmentRepository.getJobType(ChangeType.JOB.toString() + code);
 	}
 
-	public void saveJobType(JobType jobType) {
+	public void saveJobType(JobTypeDTO.SaveCode dto) {
+		JobType jobType = appointmentRepository.getJobType(ChangeType.JOB.toString() + dto.getCode());
+		
+		if (jobType == null) {
+			jobType = dto.newJobType();
+		} else {
+			jobType.changeInfo(dto.getCode()
+							  ,dto.isUseYn()
+							  ,dto.getSequence()
+							  ,dto.getComment());
+		}
+		
 		appointmentRepository.saveJobType(jobType);		
 	}
 
-	public void deleteJobType(JobType jobType) {
+	public void deleteJobType(String code) {
+		JobType jobType = appointmentRepository.getJobType(ChangeType.JOB.toString() + code);
+		
+		if (jobType == null) {
+			throw new EntityNotFoundException(code + " 엔티티가 존재하지 않습니다.");
+		}
+		
 		appointmentRepository.deleteJobType(jobType);		
 	}
